% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match_call_internal}
\alias{match_call_internal}
\title{Returns match type for each argument}
\usage{
match_call_internal(n = 1L, dots = "expand", default.formals = FALSE,
  empty.formals = FALSE, user.formals = TRUE, definition = NULL)
}
\description{
Returns match type for each argument
}
\keyword{internal}

