% Generated by roxygen2 (4.0.2): do not edit by hand
\name{match_call_old}
\alias{match_call_old}
\title{Kept here for easy reference}
\usage{
match_call_old(dots = "expand", default.formals = FALSE,
  empty.formals = FALSE, eval.formals = FALSE, user.formals = TRUE,
  parent.offset = 0L, bypass.checks = FALSE)
}
\description{
Kept here for easy reference
}
\keyword{internal}

